Prebačeni svi vanjski ključevi na tip Long umjesto Int-a

+++ app/src/main/java/hr/foi/rmai/memento/database/TaskCategoriesDAO.kt
@@ -10,7 +10,7 @@ interface TaskCategoriesDAO {
     @Query("SELECT * FROM task_categories")
     fun getAllCategories(): List<TaskCategory>
     @Query("SELECT * FROM task_categories WHERE id = :id")
-    fun getCategoryById(id: Int): TaskCategory
+    fun getCategoryById(id: Long): TaskCategory
     @Insert
     fun insertCategory(vararg category: TaskCategory)
 }
